/csr2_info "unable to open database file" error bug fix
A bug has been fixed that would return An error occurred: unable to open database file when running /csr2_info.
Core issue was that the base directory of the bot was set as the database file directory.
With that I also caught another issue that would prevent an output by referencing the wrong column name when querying for a car.