Custom Birthday Appearance
This change is purely for visually. You can now set a custom date as a birthday where the bot will change its appearance to your liking. Customization is through the .env file like always. Either set your own birthday as the birthday, the bots setup date or just an never reachable date to disable the feature.