Store Scraping Update
The bot is now able to scrape the Google play and App Store for CSR2 and CSR3 Updates in all the stores countries. Changes in either version or date will be announced to a discord channel that Server Owners can specify themselves with the command /csr2_announce_updates_add. A channel can be removed from the list with /csr2_announce_updates_delete. Similar to server channels user can choose the get notified in DMs by the bot. Commands for that are /csr2_notify_updates_add to add yourself to the list and /csr2_notify_updates_delete to remove yourself from the list again. In all cases you can choose if you want to only track CSR2, CSR3 or of course both apps, which is the default at the moment.