Add CSR2 Update Blog Website Scraping
This changes allows the bot to check https://www.csr-racing.com/csr2/csr-news for html changes. Specifically the first listed blog post on the website is being monitored for changes. Checking is every 5 minutes. To see the announcements use the same commands as for Store updates and simply change the scope.