fix non working variables tier and rarity
due to database changes to allow emoji usage for tier and rarity the search variables for those properties broke in the process. This commit fixes the described problem and slightly alters the embed display to make use of the bots emojis.
Please note that this also means in order to use the bot to its full potential you'll need to upload said emojis to your bot via the discord developer portal.
The PNGs used for the emojis can be found under: https://github.com/Nitro4CSR/CSR2-Discord-Bot-Jess/bot-emojis
Additionally you will need to replace the emoji markdown with the correct markdown of your bots emojis as the emoji_id will be different from mine.

A more central place to change the markdown codes will be created at a later time.