QOL Update to wrlist command
This update brings back page switching. Previously the interactions with the buttons just timed out due to not awaiting a function. This is now fixed and the page switching feature was also expanded by bringing page jumping and page offset jumping to life. A new button was added named "Jump to Page" that allows you to submit a value between 0 and the max pages of your wrlist. If you use mathematical signs (+ or -) offset jumping is enabled and a calculation will start to jump however many pages you entered.