improved user install experience
Due to the release of discord.py 2.5.0 it is finally possible to upload app emojis via the bot itself. This functionality was utilized here to dynamically emojify database entries for tier and rarity variables. The change was nessecary to still allow easy install and setup of the bot for beginner level self hosters.