fix zip paths and extraction paths
This issue only occurs when you run the bot with the help of a different script. A good example is running the bot on a raspi using nohup to keep it running while nobody is logged in. in that case the base dir previously was the location from where nohup was ran causing the script to never update properly. Now it should be the location of main.py file. This should now update the code correctly